Support for output using the << operator.
This patch makes it possible to call the output functions (debug, warn, error, etc)
in Log4Qt::Logger without having to specify the message immediately. The call
results in a Log4Qt::LogStream object, which implements the << operator, and
keeps track of all the data written to it. When the object gets destroyed,
all the data is sent to the appropriate method in Log4Qt::Logger.

It makes qDebug() semantics possible:
Log4Qt::Logger::logger("main")->info() << "Hello World";